requirements.txt
nltk>=3.7
numpy>=1.23.3
Pillow>=9.2.0
psaw>=0.1.0
wordcloud>=1.8.2.2
